human_pose_estimation/double_cameras/mmpose_voxelpose/requirements.txt
numpy
opencv-python
matplotlib
tqdm
pyyaml
pandas
scipy
torch
torchvision
torchaudio
openmim
